Small implemenation of a traffic light with 3
The car in principal lane should pass until a car is present on the second lane.
When the button is pressed, an even occur to allow the car on the second lane to pass, and the system return to the begin(green on principal lane). For more detail check the Readme and the pdf file.